not bababooey

+++ Collaging/src/collagefiles/model/Pixel.java
@@ -31,9 +31,10 @@ public class Pixel {
       throw new IllegalArgumentException("RGB and alpha values must be between 0 and 255");
     }
 
-    this.pixelColor = new Color(red*alpha/255,
-            green*alpha/255,
-            blue*alpha/255);
+    this.pixelColor = new Color( (int)(red* (float)alpha/255),
+
+            (int)((float)green*alpha/255),
+            (int) ((float)blue*alpha/255));
     this.red = red;
     this.green = green;
     this.blue = blue;
